last changes

+++ src/bff/operations/removeComment.js
@@ -3,15 +3,16 @@ import { sessions } from '../sessions';
 import { deleteComment } from 'bff/api/deleteComment';
 import { fetchComments } from './fetchComments';
 
-export const removeComment = async (_, commentId, postId) => {
-  // const accessRoles = [ROLE.ADMIN,];
+export const removeComment = async (hash, commentId, postId) => {
+  const accessRoles = [ROLE.ADMIN, ROLE.MODERATOR];
 
-  // if (!sessions.access(hash, accessRoles)) {
-  //   return {
-  //     error: 'Access denied',
-  //     res: null,
-  //   };
-  // }
+  const access = await sessions.access(hash, accessRoles);
+  if (!access) {
+    return {
+      error: 'Access denied',
+      res: null,
+    };
+  }
 
   await deleteComment(commentId);
   const comments = await fetchComments(postId);
